Move the methods, "update" and "fetch", to gn3.db
